[feat] Enhance GitLab mock setup and improve error handling
